backend/pyproject.toml
[project]
name = "portal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "google-genai>=1.24.0",
    "langgraph>=1.0.7",
    "typing>=3.10.0.0",
    "pydantic-settings>=2.12.0",
    "supabase>=2.27.2",
    "pypdf>=6.6.0",
    "langchain-text-splitters>=1.1.0",
    "python-multipart>=0.0.21",
]
